fix(frontend): rename routeMap to routeMapTestUtils for knip prod
knip's production pass excludes test files, so routeMap.tsx — imported only
by the generator specs — was flagged as an unused file. Rename it to match
the *TestUtils* convention knip already exempts from the production graph.
